module: Account for machine change
Building out-of-tree modules with externalsrc only runs
do_prepare_recipe_sysroot for the first MACHINE building the module. If
MACHINE changes and the module is rebuilt, the build will fail since
recipe-sysroot is not assembled and cross tools are unavailable.

This will for do_prepare_recipe_sysroot to account for MACHINE in its
sstate.
Out-of-tree modules will require EXTERNALSRC_BUILD = EXTERNALSRC (B =
S). Note that do_prepare_recipe_sysroot is still in do_configure[deps]
and do_configure, with externalsrc, already uses a stamp file to be
re-run forcibly.

It is not quite clear yet why do_prepare_recipe_sysroot is skipped in
this particular instance.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>